refactor: rename Configurator to Configuration
