+++ src/lib/ClientConstructor.ts
@@ -1,6 +1,6 @@
 import { IOAuthProvider } from 'oauth'
 
-import { CamundaPlatform8Configuration } from './Configurator'
+import { CamundaPlatform8Configuration } from './Configuration'
 
 export interface ClientConstructor {
 	config?: Partial<CamundaPlatform8Configuration>
